feat: Implement model predictive control
Cost functions are defined for smoother steering transitions along with
proper model constraints for the state update equations such that solver
models the motion model accurate enough.

MPC predicted trajectory points are appended to the solution for
visualization purposes.

Steering value is multiplied by -1 before sending it back to the server to
account for the simulator in which, a positive value implies a right turn
and a negative value implies a left turn.